third_party/WebKit/LayoutTests/fast/dom/custom/registration-context-delete-during-register-prototype-retrieval-expected.txt

Tests destroying a context during registration at the point when the prototype is retrieved.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS destroyed context
PASS caught expected exception "InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable."
PASS child done
PASS successfullyParsed is true

TEST COMPLETE
